feat(speech): comprehensive local voice support and reliability fixes
- Added 'mac' TTS backend using native 'say' command for zero-dependency macOS synthesis.
- Integrated Kokoro TTS as a high-quality local alternative (added to pyproject.toml extras).
- Implemented robust TTS fallback and health-check logic in the voice command.
- Fixed voice-id resolution to prevent cross-backend ID conflicts.
- Preset speech compute_type to 'int8' for optimized CPU inference.
- Added voice listener and speaker wrappers for cleaner audio handling.

+++ src/openjarvis/core/config.py
@@ -1205,7 +1205,7 @@ class SpeechConfig:
     model: str = "base"  # Whisper model size: tiny, base, small, medium, large-v3
     language: str = ""  # Empty = auto-detect
     device: str = "auto"  # "auto", "cpu", "cuda"
-    compute_type: str = "float16"  # "float16", "int8", "float32"
+    compute_type: str = "int8"  # "float16", "int8", "float32"
 
 
 @dataclass(slots=True)
